+++ carp_mobile_sensing/CHANGELOG.md
@@ -1,6 +1,6 @@
-## 0.40.8
+## 0.40.9
 
-* adding `pause` and `remove` life cycle methods to the `DeploymentController`
+* adding `dispose`, `pause`, and `remove` life cycle methods to the `DeploymentController`
 * support for deleting cached runtime information (deployment and task list)
 
 ## 0.40.7